[json] run serverless
