+++ package.json
@@ -15,7 +15,7 @@
     "jsonc-parser": "^2.2.1",
     "request-light": "^0.3.0",
     "vscode-json-languageservice": "^3.7.0",
-    "vscode-languageserver": "^6.1.1",
+    "vscode-languageserver": "7.0.0-next.3",
     "vscode-uri": "^2.1.2"
   },
   "devDependencies": {
